+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/main/DashboardPresenter.java
@@ -18,6 +18,7 @@ package stroom.dashboard.client.main;
 
 import com.google.gwt.dom.client.NativeEvent;
 import com.google.gwt.dom.client.Style;
+import com.google.gwt.dom.client.Style.Unit;
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.gwt.view.client.SelectionChangeEvent;
@@ -90,6 +91,7 @@ public class DashboardPresenter extends DocumentEditPresenter<DashboardView, Das
 
     private String currentParams;
     private String lastUsedQueryInfo;
+    private boolean embedded;
 
     @Inject
     public DashboardPresenter(final EventBus eventBus,
@@ -178,6 +180,11 @@ public class DashboardPresenter extends DocumentEditPresenter<DashboardView, Das
         logger.log(Level.INFO, "Dashboard Presenter setParams " + params);
 
         this.currentParams = params;
+    }
+
+    void setEmbedded(final boolean embedded) {
+        this.embedded = embedded;
+        getView().setEmbedded(embedded);
     }
 
     @Override
@@ -314,13 +321,15 @@ public class DashboardPresenter extends DocumentEditPresenter<DashboardView, Das
 
     @Override
     public void onReadOnly(final boolean readOnly) {
-        super.onReadOnly(readOnly);
+        super.onReadOnly(readOnly || embedded);
 
-        saveButton.setEnabled(isDirty() && !readOnly);
-        saveAsButton.setEnabled(true);
+        boolean enabled = !readOnly && !embedded;
 
-        addButton.setEnabled(!readOnly);
-        if (!readOnly) {
+        saveButton.setEnabled(isDirty() && enabled);
+        saveAsButton.setEnabled(enabled);
+
+        addButton.setEnabled(enabled);
+        if (enabled) {
             registerHandler(addButton.addClickHandler(event -> {
                 if ((event.getNativeButton() & NativeEvent.BUTTON_LEFT) != 0) {
                     onAdd(event);
@@ -447,6 +456,8 @@ public class DashboardPresenter extends DocumentEditPresenter<DashboardView, Das
         void setParams(String params);
 
         void setContent(View view);
+
+        void setEmbedded(boolean embedded);
     }
 
     private class AddSelectionHandler implements SelectionChangeEvent.Handler {